style : 콘서트 수정하기 UI
