+++ src/pages/HomePage.tsx
@@ -1,4 +1,5 @@
 import SearchBar from "../shared/ui/SearchBar";
+import ConcertSettingEmpty from "../features/concert/ui/ConcertSettingEmpty";
 import ConcertSetting from "../features/concert/ui/ConcertSetting";
 import TabBar from "../shared/ui/TabBar";
 
@@ -6,6 +7,7 @@ function HomePage() {
   return (
     <div className="pb-64">
       <SearchBar></SearchBar>
+      {/* <ConcertSettingEmpty></ConcertSettingEmpty> */}
       <ConcertSetting></ConcertSetting>
       <TabBar></TabBar>
     </div>
